move audit events to make sure correctness
